pyproject.toml
[project]
name = "langchain-lab"
version = "0.1.0"
description = "Laboratório para experimentar LangChain e LangGraph com PDM"
authors = [
    {name = "Eduardo Fabricio", email = "eduardo.fabricio@superbid.net"},
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0", 
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[tool.pdm.scripts]
chat = "python simple_chat_console.py"
format = "black ."
lint = "flake8 ."

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
